feat: improve custom Safe Apps' id handling
This commit fixes an issue where the custom Safe Apps' id was not being treated as an integer. By making the necessary changes, the custom Safe Apps' id is now properly converted to an integer. This ensures that the application functions correctly and avoids any potential errors related to the id type.

Co-authored-by: katspaugh <381895+katspaugh@users.noreply.github.com>